fixing report.py
sometimes I get the wrong hex address of the form `0fffff...h`. This is not `0xfff...` so it's clearly missing `x` and has `h` in the end. We can fix this by making it appropriate hex or just ignore. I'm ignoring it for now but it may come back to bite later... BEWARE